requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==38.0.4
dj-database-url==1.0.0
Django==3.2.10
django-filter==23.5
django-queryset-csv==1.1.0
django-simple-pagination==1.3
Geraldo @ https://github.com/tfs4/geraldo/archive/master.zip
gunicorn==20.1.0
idna==3.4
lxml==4.9.1
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.1.0
PySIGNFe @ https://github.com/tfs4/PySIGNFe/archive/master.zip
python-dateutil==2.8.2
python-decouple==3.1
pytz==2022.7.1
reportlab==3.6.11
requests==2.28.2
siconfipy @ git+https://github.com/tfs4/siconfipy.git@c1b1de8a1c765741261591f7a8c7e3448a77a925
signxml==3.0.0
six==1.16.0
sqlparse==0.4.3
unicodecsv==0.14.1
urllib3==1.26.14
whitenoise==6.2.0
### siops
attrs==24.2.0
chromedriver-py==127.0.6533.99
exceptiongroup==1.2.2
h11==0.14.0
outcome==1.3.0.post0
packaging==24.1
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0